+++ internal/pkg/runtime/engine/singularity/container_linux.go
@@ -82,6 +82,7 @@ type container struct {
 	skippedMount  []string
 	suidFlag      uintptr
 	devSourcePath string
+	imageBind     map[string]string
 }
 
 func create(ctx context.Context, engine *EngineOperations, rpcOps *client.RPC, pid int) error {
@@ -111,6 +112,7 @@ func create(ctx context.Context, engine *EngineOperations, rpcOps *client.RPC, p
 		mountInfoPath: fmt.Sprintf("/proc/%d/mountinfo", pid),
 		skippedMount:  make([]string, 0),
 		suidFlag:      syscall.MS_NOSUID,
+		imageBind:     make(map[string]string),
 	}
 
 	cwd := engine.EngineConfig.GetCwd()
@@ -1159,9 +1161,9 @@ func (c *container) addImageBindMount(system *mount.System) error {
 				return nil
 			})
 
-			if err := system.Points.AddBind(mount.UserbindsTag, src, destination, syscall.MS_BIND); err != nil {
-				return fmt.Errorf("while adding data bind %s -> %s: %s", src, destination, err)
-			}
+			// to honor bind ordering specified by the users
+			// we add bind mount later in addUserbindsMount
+			c.imageBind[destination] = src
 		}
 	}
 
@@ -1702,8 +1704,14 @@ func (c *container) addUserbindsMount(system *mount.System) error {
 	defaultFlags := uintptr(syscall.MS_BIND | c.suidFlag | syscall.MS_NODEV | syscall.MS_REC)
 
 	for _, b := range c.engine.EngineConfig.GetBindPath() {
-		// ignore image bind
+		// data image bind
 		if b.ID() != "" || b.ImageSrc() != "" {
+			source, ok := c.imageBind[b.Destination]
+			if ok {
+				if err := system.Points.AddBind(mount.UserbindsTag, source, b.Destination, syscall.MS_BIND); err != nil {
+					return fmt.Errorf("while adding data bind %s -> %s: %s", source, b.Destination, err)
+				}
+			}
 			continue
 		}